requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
openpyxl==3.1.5
psycopg2-binary==2.9.11
SQLAlchemy==2.0.23